only run cache workflow on the automated PRs
this isn't ideal, because in theory we want this to run before merging
*anything*, else you'll get cache misses if you time a rebuild just
wrong. but i don't know how to give it a secret and i'm afraid to leak
it. so i'll settle for just doing it for the automated PRs. at least the
majority of merges correspond to those; it's rare that a human PR even
touches the actual packages.