use rustPlatform for niri builds
crate2nix originally provided the benefit of compile times prior to the
binary cache. it no longer has a clear advantage, and also no longer
serves our needs due to not working with git-based pipewire-rs. so we're
switching to rustPlatform. the package is now more similar to the one
contained within nixpkgs.

this is also not like, the whole migration. it needs cleanup. and i'm
also going to switch the xwayland-satellite package over to rustPlatform
and get rid of crate2nix input entirely. but for now, it's going to be
in this state. it's late and i need to get up early tomorrow. and i hope
this works. it seems to work for now. maybe the CI can finally shut the
fuck up.