docs: provide a minimal hm config example
I started using Niri on a NixOS machine a few weeks ago. I found it
challenging to understand the minimal workable configuration with
home-manager + NixOS.

This adds a section to the README showing a basic configuration,
ensuring to sync pkgs version between NixOS and HM and configure
xwayland-satalite.

A little example of how to configure keybindings is also added.