add user-configured niri version to user environment
This commit reworks the method by which `niri` is included in a user's
environment. Before, it was done at the NixOS level, using
`environment.systemPackages`, meaning that all users on the system
must use the same version of `niri`.

Now, `niri` is added to the user's environment through `home.packages`;
the version added is the one set by the user in the home-manager module,
rather than at the system level.