remove unstable-note on the One Option That Sets It
Because this repo was not bleeding edge lately; there was little need to
mark these features. Most of the PRs i merged didn't, so neither did my
commits from the past ~24h.

Thank you to @wyyllou for doing it properly in #1015 though. This is the
only niri-flake addition in the 25.02..25.05 release cycle that properly
tagged the unstable option.